// frontend/profdiff-web-application/src/app/models/dto/errorMessage.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents an error message that is return within non 200 HTTP responses.  It is used in a way so that the error message is always serialized into a consistent format (JSON) by Micronaut.
 */
export interface ErrorMessage {
  /**
   * Error message to be returned in the response.
   */
  message: string;
}