// frontend/profdiff-web-application/src/app/models/dto/javaMethod.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CompilationUnitMetadata } from './compilationUnitMetadata';

/**
 * The method captured in the first profiling run (nullable).
 */
export interface JavaMethod {
  /**
   * A stable method name defined by the GraalVM compiler.
   */
  name: string;
  /**
   * List of compilation units associated with this method.
   */
  compilationUnitMetadata: Array<CompilationUnitMetadata>;
  /**
   * Total sampling period associated with a method (only if profiler file available).
   */
  totalPeriod?: number;
}