// frontend/profdiff-web-application/src/app/models/dto/warningMessage.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents a non-fatal warning that occurred during the processing or parsing of a benchmark run (runs metadata or the whole experiment with profdiff).
 */
export interface WarningMessage {
  /**
   * The human-readable description of the warning
   */
  message: string;
  /**
   * A programmatic category identifier used for UI styling or filtering
   */
  type: string;
}